Rename cert_validation to verify_steps and move functions to common_steps
