+++ e2e/namespaceissuer_steps.go
@@ -16,19 +16,19 @@ func (issCtx *IssuerContext) createNamespaceIssuer(ctx context.Context, caType s
 	return issCtx.createNamespaceIssuerInternal(ctx, caType, "")
 }
 
-func (issCtx *IssuerContext) createNamespaceIssuerWithTemplate(ctx context.Context, templateArn string, caType string) error {
-	return issCtx.createNamespaceIssuerInternal(ctx, caType, templateArn)
+func (issCtx *IssuerContext) createNamespaceIssuerWithTemplate(ctx context.Context, templateName string, caType string) error {
+	return issCtx.createNamespaceIssuerInternal(ctx, caType, templateName)
 }
 
-func (issCtx *IssuerContext) createNamespaceIssuerInternal(ctx context.Context, caType string, templateArn string) error {
+func (issCtx *IssuerContext) createNamespaceIssuerInternal(ctx context.Context, caType string, templateName string) error {
 	issCtx.issuerName = uuid.New().String() + "--namespace-issuer--" + strings.ToLower(caType)
 	issCtx.issuerType = "AWSPCAIssuer"
 	issSpec := v1beta1.AWSPCAIssuer{
 		ObjectMeta: metav1.ObjectMeta{Name: issCtx.issuerName},
 		Spec:       getIssuerSpec(caType),
 	}
-	if templateArn != "" {
-		issSpec.Spec.PCATemplateName = templateArn
+	if templateName != "" {
+		issSpec.Spec.PCATemplateName = templateName
 	}
 
 	if issCtx.secretRef != (v1beta1.AWSCredentialsSecretReference{}) {
